+++ packages/core/include/combo/types.h
@@ -12,4 +12,6 @@ typedef signed short        s16;
 typedef signed long         s32;
 typedef signed long long    s64;
 
+typedef unsigned long       uintptr_t;
+
 #endif